# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "小知，你的校园助手",
        "title": "KnowEase API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/aichat/chat/{userid}/{chatid}/": {
            "post": {
                "description": "用户发送ai请求，调用api生成回答",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "生成用户回答",
                "parameters": [
                    {
                        "description": "用户输入的问题",
                        "name": "UserMessage",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AIChatMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "生成回答成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/aichat/create": {
            "post": {
                "description": "用户创建新的聊天记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "创建新的聊天",
                "parameters": [
                    {
                        "description": "用户聊天信息",
                        "name": "AIChatHistory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AIChatHistory"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "聊天信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "创建聊天失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/aichat/{userid}/getlist": {
            "get": {
                "description": "获取用户创建的所有聊天记录列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "获取历史聊天记录列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "历史聊天记录列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "搜索记录失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/aichat/{userid}/search/{chatid}/gethistory": {
            "get": {
                "description": "获取用户某一个聊天历史里的历史聊天记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "获取历史聊天记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "UserID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "聊天ID",
                        "name": "ChatID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "历史聊天记录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "搜索记录失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/{chatid}/sseconnect": {
            "post": {
                "description": "发送请求，建立sse连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "问答"
                ],
                "summary": "流式响应连接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "聊天id",
                        "name": "chatid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "历史聊天记录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AIChatHistory": {
            "type": "object",
            "properties": {
                "chatid": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "history": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
        "models.AIChatMessage": {
            "type": "object",
            "properties": {
                "chatid": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "request": {
                    "type": "string"
                },
                "response": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}